Correction UserController.php

+++ src/Controller/UserController.php
@@ -9,8 +9,6 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Serializer\SerializerInterface;
-use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 #[Route('/api/users')]
 class UserController extends AbstractController
@@ -27,42 +25,51 @@ class UserController extends AbstractController
         $this->userRepository = $userRepository;
     }
 
-    #[Route('/all', name: 'user_list', methods: 'GET')]
-    public function getAll(UserRepository $userRepository, SerializerInterface $serializer): Response
+    #[Route('/all', name: 'user_list', methods: ['GET'])]
+    public function getAll(): Response
     {
         try {
-            $users = $userRepository->findAll();
+            $users = $this->userRepository->findAll();
+            $usernames = [];
+
             foreach ($users as $user) {
-                $usersId[] = $user->getUserName();
+                $usernames[] = $user->getUserName();
             }
-            dd($usersId);
-            return $this->json($users);
+
+            return $this->json($usernames);
         } catch (\Exception $e) {
             return $this->json(['error' => self::INTERNAL_SERVER_ERROR], 500);
         }
     }
 
-    #[Route('/new', name: 'user_new', methods: 'POST')]
+    #[Route('/new', name: 'user_new', methods: ['POST'])]
     public function new(Request $request): Response
     {
         try {
-            $data = $request->getContent();
-            $user = $data;
+            $data = json_decode($request->getContent(), true);
+
+            if (!$data) {
+                return $this->json(['error' => 'Invalid JSON'], 400);
+            }
+
+            $user = new User();
+            $user->setEmail($data['email']);
+            $user->setPassword($data['password']);
 
             $this->entityManager->persist($user);
             $this->entityManager->flush();
 
-            return $this->json($user, 201, [], ['groups' => 'user']);
+            return $this->json($user, 201);
         } catch (\Exception $e) {
             return $this->json(['error' => self::INTERNAL_SERVER_ERROR], 500);
         }
     }
 
     #[Route('/{id}', name: 'user_show', methods: ['GET'])]
-    public function show(int $id, UserRepository $userRepository): Response
+    public function show(int $id): Response
     {
         try {
-            $user = $userRepository->find($id);
+            $user = $this->userRepository->find($id);
 
             if (!$user) {
                 return $this->json(['error' => self::USER_NOT_FOUND], 404);
@@ -73,41 +80,45 @@ class UserController extends AbstractController
                 'username' => $user->getUserName(),
                 'firstName' => $user->getFirstName(),
                 'lastName' => $user->getLastName(),
-                'birthDay' => $user->getbirthday(),
+                'birthDay' => $user->getBirthday()->format('Y-m-d'),
                 'email' => $user->getEmail(),
-                'biographie' => $user->getBiographie()
+                'biographie' => $user->getBiographie(),
             ]);
         } catch (\Exception $e) {
             return $this->json(['error' => self::INTERNAL_SERVER_ERROR], 500);
         }
     }
 
-
-    #[Route('/{id}/edit', name: 'user_edit', methods: 'PUT')]
-    public function update(Request $request, User $user, SerializerInterface $serializer, ValidatorInterface $validator): Response
+    #[Route('/{id}/edit', name: 'user_edit', methods: ['PUT'])]
+    public function update(Request $request, User $user): Response
     {
         try {
-            $data = $request->getContent();
-            $serializer->deserialize($data, User::class, 'json', ['object_to_populate' => $user]);
-            $errors = $validator->validate($user);
+            $data = json_decode($request->getContent(), true);
 
-            if (count($errors) > 0) {
-                return $this->json($errors, 400);
+            if (!$data) {
+                return $this->json(['error' => 'Invalid JSON'], 400);
             }
+
+            $user->setEmail($data['email'] ?? $user->getEmail());
+            $user->setPassword($data['password'] ?? $user->getPassword());
 
             $this->entityManager->flush();
 
-            return $this->json($user, 200, [], ['groups' => 'user']);
+            return $this->json($user, 200);
         } catch (\Exception $e) {
             return $this->json(['error' => self::INTERNAL_SERVER_ERROR], 500);
         }
     }
 
-    #[Route('/{id}', name: 'user_delete', methods: 'DELETE')]
+    #[Route('/{id}', name: 'user_delete', methods: ['DELETE'])]
     public function delete(int $id): Response
     {
         try {
             $user = $this->userRepository->findOneBy(['id' => $id]);
+            if (!$user) {
+                return $this->json(['error' => self::USER_NOT_FOUND], 404);
+            }
+
             $this->entityManager->remove($user);
             $this->entityManager->flush();
